adds error handling & fixes wren-deploy calls

+++ delete_all_deployed_packages.sh
@@ -3,32 +3,43 @@
 set -e
 set -u
 
-PACKAGE_PROVIDER="jfrog"
+################################################################################
+# Includes
+################################################################################
+SCRIPT_PATH=$(dirname "${BASH_SOURCE[0]}")
 
-set -x
+source "${SCRIPT_PATH}/includes/shell_funcs.sh"
+
+################################################################################
+# Main Script
+################################################################################
+if [ ! -d "wrensec-parent" ]; then
+  echo_error "Run this from the top-level directory of Wren projects."
+  exit -1
+fi
 
 rm -rf ~/.m2/repository
 
-cd ..
+set -x
 
 cd ./wrensec-build-tools
-../wrensec-deploy-tool/wren-deploy.sh delete-all-releases "--with-provider=${PACKAGE_PROVIDER}"
+../wrensec-deploy-tool/wren-deploy.sh delete-all-releases $@
 cd ..
 
 cd ./wrensec-parent
-../wrensec-deploy-tool/wren-deploy.sh delete-all-releases "--with-provider=${PACKAGE_PROVIDER}"
+../wrensec-deploy-tool/wren-deploy.sh delete-all-releases $@
 cd ..
 
 cd ./wrensec-util
-../wrensec-deploy-tool/wren-deploy.sh delete-all-releases "--with-provider=${PACKAGE_PROVIDER}"
+../wrensec-deploy-tool/wren-deploy.sh delete-all-releases $@
 cd ..
 
 cd ./wrensec-bom
-../wrensec-deploy-tool/wren-deploy.sh delete-all-releases "--with-provider=${PACKAGE_PROVIDER}"
+../wrensec-deploy-tool/wren-deploy.sh delete-all-releases $@
 cd ..
 
 cd ./wrensec-i18n-framework
-../wrensec-deploy-tool/wren-deploy.sh delete-all-releases "--with-provider=${PACKAGE_PROVIDER}"
+../wrensec-deploy-tool/wren-deploy.sh delete-all-releases $@
 cd ..
